go back to one-branch docs
The original reason why docs were built
per branch was to allow REPL-like development
for users when writing documentation. The issue is
that when CI for images is included, then you need to
rebuild all images per board, per branch. This requires
a script that searches for every project and generates
outputs per N projects. Then this is run per M branches,
generating N * M outputs. If outputs weren't seperated by
branch, then whenever a branch would generate a website,
it could destroy the outputs another branch needs.

Ultimately I think it's too much trouble, and the markdown
syntax is intuitive enough that it's not worth the trouble.
This means mermaid diagrams will have to be made seperately
on a REPL site and copy-pasted in, but that's fine.