Servo V0 (#4)
* updated gitattributes and gitignore

* made 40v servo design

* Wrote CI for KiCad.

It currently generates PDFs and images,
but it can be extended to also make
gerbers, step files, and BOMs, and more.

* go back to one-branch docs

The original reason why docs were built
per branch was to allow REPL-like development
for users when writing documentation. The issue is
that when CI for images is included, then you need to
rebuild all images per board, per branch. This requires
a script that searches for every project and generates
outputs per N projects. Then this is run per M branches,
generating N * M outputs. If outputs weren't seperated by
branch, then whenever a branch would generate a website,
it could destroy the outputs another branch needs.

Ultimately I think it's too much trouble, and the markdown
syntax is intuitive enough that it's not worth the trouble.
This means mermaid diagrams will have to be made seperately
on a REPL site and copy-pasted in, but that's fine.

* Security fix

updated poetry dependencies because of security warning from github.

* wrote documentation

* gitignore pdfs now

* Added README