Removed Unused Functions and Fixed a Bug
Bug Fixed: If the controller sends the expected output address, then it would be allowed to passthrough the translator. 

This is unintended behavior, as it would stop the controller from being able to individually write different configurations to identical sensors positioned before and after the translator. The fix mangles the address so that the output device will not be addressed unless it is intended.